Add potential Meeple sprites
Added 3 knight sprites and 3 orc sprites that could potentially serve as meeple sprites. My idea was to use the 'head' sprites as visuals for how many meeples you can place, and the full sprites as visuals for placed meeples.
These can be downloaded and implemented into other branches if necessary, or replaced with better alternative sprites.